Get available palette names and series

c4a_palettes lists all available cols4all color palettes. Palettes are organized by series. The available series are listed with c4a_series. Palettes are also organized per functional type, where we currently support: categorical "cat", sequential "seq", and diverging "div"" palette types. The function c4a_types lists all available types. The function c4a_overview gives an overview table of the number of palette per series and type. In an IDE with auto-completion (such as RStudio) it is possible to browse through the palette names with .P (using $ like in lists).

Usage

c4a_palettes(
  type = c("all", "cat", "seq", "div"),
  series = NULL,
  full.names = TRUE
)

c4a_series(type = c("all", "cat", "seq", "div"), as.data.frame = TRUE)

c4a_types(series = NULL, as.data.frame = TRUE)

c4a_overview()

.P

Format

An object of class environment of length 17.

Arguments

type

type of color palette: one of "all" (all palettes), "cat" (categorical/qualitative palettes), "seq" (sequential palettes) and "div" (diverging palettes).

series

series to list the palettes from. Run c4a_series to see the options.

full.names

should full names, i.e. with the prefix "series."? By default TRUE.

as.data.frame

should c4a_series and c4a_types return the result as a data.frame, with description included as a column?

Value

names of the loaded color palettes

See also

Examples

c4a_series()
#>       series                                         description
#> 1     brewer                                ColorBrewer palettes
#> 2        c4a                  cols4all palettes (in development)
#> 3      carto                          Palettes designed by CARTO
#> 4        hcl  Palettes from the Hue Chroma Luminance color space
#> 5     kovesi                   Palettes designed by Peter Kovesi
#> 6        met Palettes inspired by The Metropolitan Museum of Art
#> 7       misc                              Miscellaneous palettes
#> 8      miscs                                                <NA>
#> 9      parks                 Palettes inspired by National Parks
#> 10 pinkfloyd     Palettes extracted from Pink Floyd album covers
#> 11      poly               Qualitative palettes with many colors
#> 12     scico     Scientific colour map palettes by Fabio Crameri
#> 13   seaborn            Palettes from the Python library Seaborn
#> 14   stevens                Bivariate palettes by Joshua Stevens
#> 15   tableau                        Palettes designed by Tableau
#> 16       tol                       Palettes designed by Paul Tol
#> 17   viridis          Palettes fom the Python library matplotlib
#> 18       wes                   Palettes from Wes Anderson movies

c4a_types()
#>   type                          description
#> 1  cat                          categorical
#> 2  seq                           sequential
#> 3  div                            diverging
#> 4 bivs  bivariate (sequential x sequential)
#> 5 bivc bivariate (sequential x categorical)
#> 6 bivd   bivariate (sequential x diverging)
#> 7 bivg bivariate (sequential x desaturated)

c4a_overview()
#>           cat seq div bivs bivc bivd bivg
#> brewer      9  18   9    2    1    1   NA
#> c4a        NA  NA   2    2   NA    2    5
#> carto       6  21   7   NA   NA   NA   NA
#> hcl         9  23  11   NA   NA   NA   NA
#> kovesi     NA  17  13   NA   NA   NA   NA
#> met        33   8  14   NA    1   NA   NA
#> misc        1  NA  NA   NA    3   NA   NA
#> miscs       4  NA  NA   NA   NA   NA   NA
#> parks      22   5   3   NA   NA   NA   NA
#> pinkfloyd  16  NA  NA   NA   NA   NA   NA
#> poly        9  NA  NA   NA   NA   NA   NA
#> scico      21  21  10   NA    2   NA    1
#> seaborn     6   4   2   NA   NA   NA   NA
#> stevens    NA  NA  NA    5   NA   NA   NA
#> tableau    29  23  28   NA   NA   NA   NA
#> tol         8   8   4   NA   NA   NA   NA
#> viridis    NA   7   1   NA   NA   NA   NA
#> wes        23  NA   1   NA   NA   NA   NA

c4a_palettes(type = "cat", series = "tol")
#> [1] "tol.bright"   "tol.contrast" "tol.vibrant"  "tol.muted"    "tol.medium"  
#> [6] "tol.light"    "tol.dark"     "tol.rainbow" 

c4a_palettes(type = "seq", series = "kovesi")
#>  [1] "kovesi.linear_grey"          "kovesi.rainbow_bu_rd"       
#>  [3] "kovesi.rainbow_bu_pk"        "kovesi.linear_ternary_blue" 
#>  [5] "kovesi.linear_ternary_green" "kovesi.linear_ternary_red"  
#>  [7] "kovesi.linear_yl_rd_bk"      "kovesi.linear_wh_rd_bk"     
#>  [9] "kovesi.linear_green"         "kovesi.linear_yl_mg_bu"     
#> [11] "kovesi.linear_wh_mg_bu"      "kovesi.linear_blue"         
#> [13] "kovesi.linear_tq_bu"         "kovesi.linear_wh_yl_gn_bu"  
#> [15] "kovesi.linear_yl_gn_bu"      "kovesi.isoluminant_tq_or"   
#> [17] "kovesi.linear_terrain"      

# handy when auto-completion is available:
.P$kovesi$seq$linear_terrain
#> [1] "kovesi.linear_terrain"