cols4all overview

cols4all stands for: color palettes for all people, including those with color vision deficiency. Popular color palette series, such as ColorBrewer, have been organized by type and have been scored on several properties such as color-blind-friendliness and fairness (i.e. do colors stand out equally?). Own palettes can also be loaded and analysed. Besides the common palette types (categorical, sequential, and diverging) it also includes bivariate color palettes. ggplot2 scales are included.

Details

This page provides a brief overview of all package functions.

Main functions

c4a_gui	Dashboard for analyzing the palettes

c4a	Get the colors from a palette (c4a_na for the associated color for missing values)

c4a_plot	Plot a color palette

Palette names and properties

c4a_palettes	Get available palette names

c4a_series	Get available series names

c4a_overview	Get an overview of palettes per series x type

c4a_citation	Show how to cites palettes (with bibtex code)

c4a_info	Get information from a palette, such as type and maximum number of colors)

.P	Environment via which palette names can be browsed with auto-completion (using $)

Importing and exporting palettes

c4a_data	Build color palette data

c4a_load	Load color palette data

c4a_sysdata_import	Import system data

c4a_sysdata_export	Export system data

See also

Author

Maintainer: Martijn Tennekes mtennekes@gmail.com

Other contributors:

Marco Puts mputs@acm.org [contributor]

Achim Zeileis Achim.Zeileis@R-project.org [contributor]

Jakub Nowosad nowosad.jakub@gmail.com [contributor]

Robin Lovelace rob00x@gmail.com [contributor]

Helgasoft contact@helgasoft.com [contributor]

Matthew Petroff matthew@mpetroff.net [contributor]

Olivier Roy [contributor]