Set cols4all options

Get or set global options for c4a. Works similar as the base function options

Usage

c4a_options(...)

Arguments

...

Use character values to retrieve options. To set options, either use named arguments (where the names refer to the options), a list that consists of those options.

Value

A list of options

Details

Option	Description

defaults	Default palettes per type

CBF_th	Parameters that label a palette as color blind friendly

CBU_th	Parameters that label a palette as color blind unfriendly

CrangeFair	Maximum chroma range for which a palette is considered harmonic

CrangeUnfair	Minimum chroma range for which a palette is considered disharmonic

LrangeFair	Maximum luminance range for which a palette is considered harmonic

LrangeUnfair	Minimum luminance range for which a palette is considered disharmonic

Cintense	Chroma of colors that are considered intense

Cpastel	Chroma of colors that are considered 'pastel'

HwidthDivRainbow	A diverging palette is labeled as 'rainbow hue' if HwidthL or HwidthR are at least HwidthDivRainbow

HwidthDivSingle	A diverging palette is labeled as 'single hue' if HwidthL and HwidthR are at most HwidthDivSingle

HwidthSeqRainbow	A sequential palette is labeled as 'rainbow hue' if Hwidth is at least HwidthSeqRainbow

HwidthSeqSingle	A sequential palette is labeled as 'single hue' if Hwidth is at most HwidthSeqSingle